refactor(legolas, gandalf): collapse to interface's native surfaces
Design-review follow-up to f825e3f. The first pass invented a fourth
surface on top of IPlayerAreaState's three (inject / query .CurrentArea /
Subscribe): a local _cachedArea field in PlayerLogIngestionService, and
in Gandalf an UpdateCurrentArea(string?) push API on LootSource plus a
LootIngestionService Subscribe that did nothing but forward into that
push API. Drop all three; each consumer now uses only the native
surfaces.

Legolas PlayerLogIngestionService:
* Drop _cachedArea field. OnAreaChanged simplifies to
  `if (evt.Current is not null) _areaCalibration.SelectArea(evt.Current);`
  — the latch + null-reset that the field was supporting only mattered
  pre-#789 when ApplyAreaIfChanged had two callers (StartAsync seed +
  every per-line dispatch). Now Subscribe is the single caller and the
  upstream contract guarantees no consecutive same-area events (Snapshot
  fires once on attach; Changed fires only on genuine transitions).

Gandalf LootSource:
* Take `IPlayerAreaState? areaState` directly (back to where it was
  before f825e3f bridged through a push API). Drop _cachedArea field and
  UpdateCurrentArea method. Read `_areaState?.CurrentArea` at the two
  chest-commit sites (OnChestInteraction + the cow-rejection
  anchor-from-rejection branch in OnChestCooldownObserved).

Gandalf LootIngestionService:
* Drop the IPlayerAreaState ctor param, _areaState field, _areaSub
  field, StartAsync / StopAsync overrides, and OnAreaChanged method.
  After the Observe push-in retired in f825e3f this service had no area
  work to do; the Subscribe-and-forward bridge was pure plumbing for the
  push API.

GandalfModule DI:
* LootSource registration takes `areaState: sp.GetService<IPlayerAreaState>()`.
* LootIngestionService is back to its pre-#789 signature.

Tests:
* FakePlayerAreaState hoisted out of LootIngestionServiceTests into
  tests/Gandalf.Tests/FakePlayerAreaState.cs (internal, project-scoped)
  so LootSourceTests can use it too.
* LootSourceTests.Build returns the FakePlayerAreaState as a tuple slot;
  the four area-driven tests (stamps-current-area, sticky-once-known,
  stamps-late, area-scoped friendly-name + region) call
  areaState.SetArea(...) instead of src.UpdateCurrentArea(...).
* LootIngestionServiceTests.Harness wires the FakePlayerAreaState
  into LootSource (not the service) and drops the FakePlayerAreaState
  nested class that lived in this file.

Verification: dotnet build Mithril.slnx clean (0/0); dotnet test
Mithril.slnx clean; grep --include='*.cs' across both modules for
PlayerAreaTracker / UpdateCurrentArea / _cachedArea returns no hits.
