+++ src/Gandalf.Module/Services/LootIngestionService.cs
@@ -1,5 +1,4 @@
 using Gandalf.Parsing;
-using Mithril.GameState.Areas;
 using Microsoft.Extensions.Hosting;
 using Mithril.Shared.Diagnostics;
 using Mithril.Shared.Logging;
@@ -19,20 +18,12 @@ namespace Gandalf.Services;
 /// anchor (the <see cref="DefeatCooldownParser"/> rejection text remains as
 /// a diagnostic-only "cooldown still active" observation).
 ///
-/// <para><b>Area→chest-stamp bridge (#789).</b> Subscribes to
-/// <see cref="IPlayerAreaState"/> on <c>StartAsync</c> and forwards every
-/// notification to <see cref="LootSource.UpdateCurrentArea"/> so a chest
-/// commit (<see cref="LootBracketTracker"/> → <see cref="LootSource.OnChestInteraction"/>)
-/// stamps <c>LearnedChest.Area</c> with the area the player was in at
-/// commit time (#178). The Snapshot replay synthesised by
-/// <see cref="IPlayerAreaState.Subscribe"/> on attach delivers the seed
-/// area before any chest interaction can reach the bracket tracker — so
-/// session-start commits stamp the correct area without a separate eager
-/// read of <see cref="IPlayerAreaState.CurrentArea"/>. Subscribing here
-/// (NOT via <c>IPlayerWorld.Bus</c>) is load-bearing while the eager
-/// pre-warm in <c>PlayerAreaWorldRegistration</c> is still active — its
-/// retirement is the follow-on PR once both consumer migrations land
-/// (#769).</para>
+/// <para><b>Area→chest-stamp bridge (#790).</b> Owned by <see cref="LootSource"/>
+/// directly — it injects <c>IPlayerAreaState</c> and queries
+/// <c>CurrentArea</c> at chest-commit time. This service no longer has any
+/// area dep; the per-line area-tracker push-in retired in #790 because
+/// the producer's L1 subscription already owns the <c>LOADING LEVEL</c>
+/// envelope path end-to-end.</para>
 ///
 /// <para><b>L1 migration disposition (#549 Gandalf row, #550 PR 3 archetype-B).</b>
 /// <list type="bullet">
@@ -76,11 +67,9 @@ public sealed class LootIngestionService : BackgroundService
     private readonly LootBracketTracker _bracket;
     private readonly BossKillCreditParser _bossKill;
     private readonly DefeatCooldownParser _defeatCooldown;
-    private readonly IPlayerAreaState _areaState;
     private readonly LootSource _source;
     private readonly IDiagnosticsSink? _diag;
     private ILogSubscription? _subscription;
-    private IDisposable? _areaSub;
     private bool _firstObservationLogged;
 
     public LootIngestionService(
@@ -88,7 +77,6 @@ public sealed class LootIngestionService : BackgroundService
         LootBracketTracker bracket,
         BossKillCreditParser bossKill,
         DefeatCooldownParser defeatCooldown,
-        IPlayerAreaState areaState,
         LootSource source,
         IDiagnosticsSink? diag = null)
     {
@@ -96,29 +84,8 @@ public sealed class LootIngestionService : BackgroundService
         _bracket = bracket;
         _bossKill = bossKill;
         _defeatCooldown = defeatCooldown;
-        _areaState = areaState ?? throw new ArgumentNullException(nameof(areaState));
         _source = source;
         _diag = diag;
-    }
-
-    public override Task StartAsync(CancellationToken cancellationToken)
-    {
-        // Area bridge first — Subscribe fires a synthetic Snapshot inside
-        // the call carrying the seed area, so LootSource sees the correct
-        // area before any chest commit can reach it from the L1 subscription
-        // below. Subscribing here (NOT via IPlayerWorld.Bus) is load-bearing
-        // while the eager pre-warm in PlayerAreaWorldRegistration is still
-        // active — bus subscribers never see the seed because that
-        // transition is applied directly to the folder and bypasses the bus.
-        _areaSub = _areaState.Subscribe(OnAreaChanged);
-        return base.StartAsync(cancellationToken);
-    }
-
-    public override async Task StopAsync(CancellationToken cancellationToken)
-    {
-        _areaSub?.Dispose();
-        _areaSub = null;
-        await base.StopAsync(cancellationToken).ConfigureAwait(false);
     }
 
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
@@ -153,20 +120,8 @@ public sealed class LootIngestionService : BackgroundService
     {
         _subscription?.Dispose();
         _subscription = null;
-        _areaSub?.Dispose();
-        _areaSub = null;
         base.Dispose();
     }
-
-    /// <summary>
-    /// Forward area notifications to <see cref="LootSource"/>. Both
-    /// <see cref="PlayerAreaChangeKind.Snapshot"/> (seed on attach) and
-    /// <see cref="PlayerAreaChangeKind.Changed"/> (live transition) carry
-    /// only the latest <see cref="PlayerAreaChanged.Current"/> we care
-    /// about for stamping chest commits.
-    /// </summary>
-    private void OnAreaChanged(PlayerAreaChanged evt) =>
-        _source.UpdateCurrentArea(evt.Current);
 
     private void Dispatch(LocalPlayerLogLine payload)
     {
